Patch 2: warning-fix.patch - Using the prefix version of ++ operator, it is more efficient once it does not create temporary object - Assignment operator must not return const - Removed the extra qualification 'dns_hasher::', because static methods can call just other static methods. Some compilers even consider this as an error. - Some member variables were not initialized in constructor, or the initialization was in the constructor body instead of initialize list. - Some parameters were shadowing other identifiers. - Commented unused parameter. - Explicit cast between uint16 and int literals and variables.
Originally from guigasurf

+++ boost/net/network_array.hpp
@@ -318,7 +318,7 @@ public:
       position(p);
     }
 
-    *((uint32_t *)&array<uint8_t,N>::elems[nap]) = htonl(d.to_ulong());
+    *((uint32_t *)&array<uint8_t,N>::elems[nap]) = static_cast<uint32_t>(htonl(d.to_ulong()));
 
     if( incpos )
     {
